refactor: [comment] Response(controller), Request(domain) 분리
