+++ src/main/java/site/travellaboratory/be/comment/domain/request/CommentUpdateRequest.java
@@ -1,0 +1,7 @@
+package site.travellaboratory.be.comment.domain.request;
+
+public record CommentUpdateRequest(
+    String replyComment
+) {
+
+}
