--- src/main/java/site/travellaboratory/be/comment/presentation/response/writer/CommentUpdateRequest.java
@@ -1,7 +1,0 @@
-package site.travellaboratory.be.comment.presentation.response.writer;
-
-public record CommentUpdateRequest(
-    String replyComment
-) {
-
-}
